+++ peeringdb/_update.py
@@ -5,18 +5,18 @@ import logging
 import threading
 from itertools import chain
 from datetime import datetime
+from contextlib import contextmanager
 
 from peeringdb import resource, get_backend
 from peeringdb import _sync, _fetch, _config_logs
+from peeringdb.resource import all_resources
 
-try:
-    from peeringdb import _tasks_async as _tasks
-except:
-    from peeringdb import _tasks_sequential as _tasks
+from peeringdb import _tasks_sequential as _tasks
+wrap_generator = _tasks.wrap_generator
 
 
 class _CancelSave(Exception):
-    "Dummy exception to cancel transaction"
+    "Token exception to cancel transaction"
     pass
 
 
@@ -57,25 +57,28 @@ class Updater(object):
 
     def update_all(self, rs=None, since=None):
         "Sync all objects for the relations rs (if None, sync all resources)"
-        self._log.info("Updating resources: %s", ' '.join(r.tag for r in rs))
+        self._log.info("Updating resources: %s", ' '.join(r.tag for r in (rs or [])))
 
         if rs is None:
             rs = resource.all_resources()
         ctx = self._ContextClass(self)
         for r in rs:
-            self._atomic_update(lambda: ctx.sync_resource(r, since=since))
+            with self._transaction():
+                ctx.sync_resource(r, since=since)
 
     def _update(self, res, fetch_func, depth):
         ctx = self._ContextClass(self)
         data, e = fetch_func()
         if e: raise e
         self._log.info("Updates to be processed: {}".format(len(data)))
-        self._atomic_update(lambda: ctx.sync_rows(res, data, depth + 1))
+        with self._transaction():
+            ctx.sync_rows(res, data, depth + 1)
 
-    def _atomic_update(self, sync_func):
+    @contextmanager
+    def _transaction(self):
         try:
             with get_backend().atomic_transaction():
-                sync_func()
+                yield
                 if self.dry_run:
                     raise _CancelSave
                 self._log.debug('Committing transaction')
@@ -113,10 +116,8 @@ class UpdateContext(object):
         return _tasks.gather(self._schedule_rows(res, rows, depth))
 
     def _schedule_rows(self, res, rows, depth):
-        return [
-            self.set_job((res, row['id']), self.sync_row,
-                         (res, row, depth)) for row in rows
-        ]
+        return [self.set_job((res, row['id']), self.sync_row, (res, row, depth))
+                for row in rows]
 
     def get_task(self, key):
         """Get a scheduled task, or none"""
@@ -147,7 +148,7 @@ class UpdateContext(object):
                         'new task' if not had else 'dup')
         return job
 
-    def pending_tasks(self, res):
+    def pending_jobs(self, res):
         "Synchronized access to tasks"
         jobs, lock = self._jobs
         with lock:
@@ -170,40 +171,37 @@ class UpdateContext(object):
             self._log.info('Fetched no data for %s-%s', res.tag, pk)
             data, e = self.fetcher.fetch_deleted(res, pk, 0)
             if not data:
-                self._log.info(
-                    'Fetched no deleted objects for %s-%s, aborting', res.tag,
-                    pk)
+                self._log.info('Fetched no deleted objects for %s-%s, aborting',
+                               res.tag, pk)
                 return
             row = data[0]
         yield self.sync_row(res, row, depth)
 
     @_tasks.wrap_generator
     def sync_row(self, res, row, depth):
-        self._log.debug("sync_row(%s, %s, %s)", res, row['id'], depth)
-        if self.disable_partial and depth > 0:
-            raise ValueError('depth > 0 sync is disabled')
-
         B = get_backend()
-        obj, fetched, dangling = _sync.initialize_object(B, res, row)
-
-        # self._log.debug(' fetched: %s', fetched)
-        # self._log.debug(' dangling: %s', dangling)
-
         def _have(R, pks):
             have = B.get_objects(B.get_concrete(R), pks)
             return set(have.values_list('id', flat=True))
 
+        self._log.debug("sync_row(%s, %s, %s)", res.tag, row['id'], depth)
+        if self.disable_partial and depth > 0:
+            raise ValueError('depth > 0 sync is disabled')
+
         # Before attempting to set the related-object fields, ensure they are synced
         # Skip all ref'd objects that we already have, or have scheduled to update
+        fetched, dangling = _sync.extract_relations(B, res, row)
         sync_jobs = []
+
         for R, sub in fetched.items():
             have = _have(R, set(sub.keys()))
             sync_jobs.extend(
                 self.set_job((R, pk), self.sync_row, (R, subrow, depth - 1))
                 for pk, subrow in sub.items() if not (pk in have))
+
         for R, pks in dangling.items():
             pks = pks.difference(_have(R, pks))
-            pending = self.pending_tasks(R)
+            pending = self.pending_jobs(R)
             needpks = []
             for pk in pks:
                 if pk in pending:
@@ -213,55 +211,37 @@ class UpdateContext(object):
             if not needpks:
                 continue
 
-            def fetch(_R=R, _pks=needpks):
-                return self.fetcher.fetch_all_latest(_R, 0, dict(id__in=_pks))
-
+            def fetch_dangling(_R=R, _pks=needpks):
+                return self.fetcher.fetch_all(_R, 0, dict(id__in=_pks))
             fetch_job = _tasks.UpdateTask(
-                self.fetch_and_index(fetch), (R, None))
-
+                self.fetch_and_index(fetch_dangling), (R, None))
             sync_jobs.extend(
                 self.set_job((R, pk), self.update_after,
                              (R, pk, depth - 1, fetch_job)) for pk in pks)
 
+        obj = _sync.initialize_object(B, res, row)
+        _sync.set_scalars(B, res, obj, row)
+
+        # Resolve refs and then save full object
+        self._log.debug(' waiting for: %s', sync_jobs)
+        yield _tasks.gather(sync_jobs)
+        self._log.debug("sync_row(%s, %s, %s) (resumed)", res.tag, row['id'], depth)
+
+        _sync.set_single_relations(B, res, obj, row)
         # Detect integrity gaps
         dup_fields, missing = _sync.clean_helper(B, obj, B.clean)
 
         if dup_fields:
-            # For non-uniques, try to update conflicting object
-            field = dup_fields[0]
-            value = getattr(obj, field)
-            try:
-                dup = B.get_object_by(B.get_concrete(res), field, value)
-                dup_id = dup.id
-                dup.delete(hard=True)
-            except B.object_missing_error(B.get_concrete(res)):  # shouldn't happen
-                raise Exception('internal error')
-
-            self._log.debug('dup: %s = %s', field, value)
-
-            def fetch():
-                data, err = self.fetcher.fetch_latest(res, dup_id, 0)
-                if isinstance(err, _fetch.NotFoundException):
-                    # case where the local duplicate has been deleted
-                    return {}, None
-                return data, err
-
-            fetch_job = _tasks.UpdateTask(
-                self.fetch_and_index(fetch),
-                (res, dup_id))
-            job = self.set_job((res, dup_id), self.update_after,
-                               (res, dup_id, 0, fetch_job))
+            job = self._handle_duplicate(res, obj, dup_fields)
             sync_jobs.append(job)
 
-        if missing:
-            # ignore expected missing refs
-            for R, pks in missing.items():
-                for pk in pks:
-                    if pk not in dangling[R]:  # shouldn't happen
-                        raise RuntimeError("Unexpected missing relation",
-                                           (R, pk))
+        # ignore expected missing refs
+        for R, pks in missing.items():
+            for pk in pks:
+                if pk not in dangling[R]:  # shouldn't happen
+                    raise RuntimeError("Unexpected missing relation",
+                                       (R, pk))
 
-        # Preliminary save so this object exists for its dependencies
         _sync.patch_object(B, res, obj, self.updater.strip_tz)
 
         # Ignore new objects for consistency - TODO: test
@@ -270,19 +250,52 @@ class UpdateContext(object):
                            res.tag, row['id'])
             return
 
+        # Preliminary save so this object exists for its dependents
+        # XXX Doesn't work in MySQL, single-refs must exist
         if not self.updater.dry_run:
             B.save(obj)
+            # B.get_objects(res, pk=row['id']).update()
 
-        # Now, resolve refs and then save full object
-        self._log.debug(' waiting for (%s-%s): %s', res.tag, row['id'],
-                        sync_jobs)
+        def finish():
+            _sync.set_single_relations(B, res, obj, row)
+            _sync.set_many_relations(B, res, obj, row)
+
+            if self.updater.dry_run: return
+            B.clean(obj)
+            B.save(obj)
+        self.set_job((res, row['id']), self.sync_row_finish,
+                     (res, row['id'], finish, sync_jobs))
+
+
+    @_tasks.wrap_generator
+    def sync_row_finish(self, res, pk, func, sync_jobs):
+        # Resolve refs and then save full object
+        self._log.debug(' waiting for: %s', sync_jobs)
         yield _tasks.gather(sync_jobs)
-        self._log.debug("sync_row(%s, %s, %s) (resumed)", res.tag, row['id'],
-                        depth)
+        self._log.debug("sync_row(%s, %s) (finish)", res.tag, pk)
+        func()
 
-        if self.updater.dry_run:
-            return
 
-        _sync.set_object_relations(B, res, obj, row)
-        B.clean(obj)
-        B.save(obj)
+    def _handle_duplicate(self, res, obj, dup_fields):
+        B = get_backend()
+        # Try to update conflicting object
+        field = dup_fields[0]
+        value = getattr(obj, field)
+        try:
+            dup = B.get_object_by(B.get_concrete(res), field, value)
+            dup_id = dup.id
+            dup.delete(hard=True)
+        except B.object_missing_error(B.get_concrete(res)):  # shouldn't happen
+            raise Exception('internal error')
+
+        self._log.debug('dup: %s = %s', field, value)
+
+        def fetch_dup():
+            data, err = self.fetcher.fetch_latest(res, dup_id, 0)
+            if isinstance(err, _fetch.NotFoundException):
+                # case where the local duplicate has been deleted
+                return {}, None
+            return data, err
+
+        fetch_job = _tasks.UpdateTask(self.fetch_and_index(fetch_dup), (res, dup_id))
+        return self.set_job((res, dup_id), self.update_after, (res, dup_id, 0, fetch_job))