+++ tests/test_client.py
@@ -4,8 +4,7 @@ import helper
 import peeringdb
 from peeringdb import resource
 
-helper.set_data_path(__file__, 'data')
-client = helper.client_fixture('insert_full.sql')
+client = helper.client_fixture('full')
 NET0 = 7
 
 def test_nonexistent_config():