+++ tests/test_sync.py
@@ -22,12 +22,13 @@ def get_pclient():
     c._updater._ContextClass = _PartialEnabledContext
     return c
 
-client_dup = helper.client_fixture('insert_nonunique.sql')
+client_dup = helper.client_fixture('full_nonunique')
 
 def test_full(client_empty):
     client = get_client()
     rs = all_resources()
     client.update_all(rs)
+    assert client.get(Network, FIRST_NET)
 
 # Ensure fetching is not order-dependent
 def test_reversed(client_empty):